refactor: Use symlink_install and move symlink creation to install step
Usage of `symlink_install` is recommended by Brew Formula Cookbook, see
https://docs.brew.sh/Formula-Cookbook#file-level-operations.

This allow to use of `bin/emacs --version`, when `--with-starter` has been specified. The latter has changed Emacs to
require `native-lisp` directory when called with `--version`.

Also add some sanity checks for source and target when creating a symlink, thus allowing to skip a check for
native compilation.